refactor interface in 3.gRPC
