+++ course4/1.RPC/3.gRPC/task4.1.3.1/internal/controller/responder.go
@@ -8,15 +8,6 @@ import (
 	"net/http"
 )
 
-type Responder interface {
-	OutputJSON(w http.ResponseWriter, data any)
-
-	ErrorUnAuthorized(w http.ResponseWriter, err error)
-	ErrorBadRequest(w http.ResponseWriter, err error)
-	ErrorInternal(w http.ResponseWriter, err error)
-	ErrorForbidden(w http.ResponseWriter, err error)
-}
-
 type Response struct {
 	Success bool   `json:"success"`
 	Data    any    `json:"data"`
